Added bullet attack

+++ Assets/Scripts/Combat/PlayerCombatManager.cs
@@ -87,14 +87,14 @@ public class PlayerCombatManager : MonoBehaviour {
         }
     }
 
-    public void ReceivePlayerAttack(Vector3 _AttackVelocity, int _AttackTimeSpan) {
+    public void ReceivePlayerAttack(Vector3 _AttackVelocity, float _AttackTimeSpan) {
         PhotonView _view = PhotonView.Get(this);
         Debug.Log("Invoking OnAttacked on MasterClient");
         _view.RPC("OnPlayerAttacked", RpcTarget.All, _AttackVelocity.x, _AttackVelocity.y, _AttackVelocity.z, _AttackTimeSpan);
     }
 
     [PunRPC]
-    public IEnumerator OnPlayerAttacked(float _x, float _y, float _z, int _time) {
+    public IEnumerator OnPlayerAttacked(float _x, float _y, float _z, float _time) {
         Debug.Log("OnAttacked executed with " + _x + " " + _y + " " + _z + " with knockback " + _time);
 
         Handler.enabled = false;
